LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/radicals/glyph-level-mirroring-expected.txt


PASS LTR Direction
FAIL RTL Direction assert_approx_equals: Base glyph expected 32 +/- 1 but got 80
